Create 2 sources files for the C readable implementation
