+++ src/unix/lwt_unix_windows.h
@@ -538,7 +538,6 @@ LWT_NOT_AVAILABLE6(unix_madvise)
 LWT_NOT_AVAILABLE4(unix_mincore)
 LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
 LWT_NOT_AVAILABLE1(unix_guess_blocking_job)
-LWT_NOT_AVAILABLE1(unix_readable)
 LWT_NOT_AVAILABLE1(unix_writable)
 LWT_NOT_AVAILABLE1(unix_open_job)
 LWT_NOT_AVAILABLE1(unix_stat_job)
