+++ src/unix/windows_c/windows_readable.c
@@ -1,0 +1,4 @@
+#if defined(LWT_ON_WINDOWS)
+#include "lwt_unix.h"
+LWT_NOT_AVAILABLE1(unix_readable)
+#endif
